+++ src/plugin/js/index.ts
@@ -1,28 +1,25 @@
-
 import "../css/index.scss";
 
-import type { Api } from 'reveal.js';
+import type { Api } from "reveal.js";
+import { pluginDebug as debug, PluginBase, pluginCSS } from "reveal.js-plugintoolkit";
 import type { Config } from "./config";
-import { defaultConfig } from './config.js'
-import { PluginBase, pluginCSS, pluginDebug as debug } from 'reveal.js-plugintoolkit';
-import { main } from './main';
+import { defaultConfig } from "./config.js";
+import { main } from "./main";
 
-
-const PLUGIN_ID = 'copycode';
+const PLUGIN_ID = "copycode";
 
 const init = async (plugin: PluginBase<Config>, deck: Api, config: Config): Promise<void> => {
+	if (debug && config.debug) {
+		debug.initialize(true, PLUGIN_ID);
+	}
 
-    if (debug && config.debug) {
-        debug.initialize(true, PLUGIN_ID);
-    }
-
-    await pluginCSS(plugin, config);
-    await main(deck, config);
-}
-
-export default () => {
-    const plugin = new PluginBase(PLUGIN_ID, init, defaultConfig);
-    return plugin.createInterface();
+	await pluginCSS(plugin, config);
+	await main(deck, config);
 };
 
-export type { Config } from './config';
+export default () => {
+	const plugin = new PluginBase(PLUGIN_ID, init, defaultConfig);
+	return plugin.createInterface();
+};
+
+export type { Config } from "./config";